libs/report_writer/pyproject.toml
[tool.poetry]
name = "report_writer"
version = "0.1.0"
description = ""
authors = ["renatormc <renatomartinsrmc@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
jinja2 = "*"
docxtpl = "*"
beautifulsoup4 = "*"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]
mypy = "*"
types-setuptools = "*"
pytest = "*"
types-beautifulsoup4 = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
